Rename primary ruler window
